Update TS directory structure
