+++ web/ts/components/header.ts
@@ -1,0 +1,15 @@
+import { Notifications } from "../notifications";
+import { ToggleableElement } from "../utilities/ToggleableElement";
+
+export function header() {
+    return {
+        userContext: new ToggleableElement(new Map([["themePicker", new ToggleableElement()]])),
+
+        notifications: new Notifications(),
+        notification: new ToggleableElement(),
+        toggleNotification(set?: boolean) {
+            this.notification.toggle(set);
+            this.notifications.writeToStorage(true);
+        },
+    };
+}
