+++ web/ts/entry/home.ts
@@ -1,1 +1,2 @@
 export * from "../views/home";
+export * from "../components/header";
