+++ web/ts/utilities/ToggleableElement.ts
@@ -1,0 +1,21 @@
+export class ToggleableElement {
+    private readonly children: Map<string, ToggleableElement>;
+
+    public value: boolean;
+
+    constructor(children?: Map<string, ToggleableElement>, value = false) {
+        this.children = children || new Map<string, ToggleableElement>();
+        this.value = value;
+    }
+
+    getChild(name: string): ToggleableElement {
+        return this.children.get(name);
+    }
+
+    toggle(set?: boolean) {
+        this.value = set ?? !this.value;
+        if (!this.value) {
+            this.children.forEach((c) => c.toggle(false));
+        }
+    }
+}
